+++ sqlglot/dialects/bigquery.py
@@ -13,6 +13,7 @@ from sqlglot.dialects.dialect import (
     max_or_greatest,
     min_or_least,
     no_ilike_sql,
+    parse_date_delta_with_interval,
     rename_func,
     timestrtotime_sql,
     ts_or_ds_to_date_sql,
@@ -21,25 +22,6 @@ from sqlglot.helper import seq_get
 from sqlglot.tokens import TokenType
 
 E = t.TypeVar("E", bound=exp.Expression)
-
-
-def date_add(expression_class: t.Type[E]) -> t.Callable[[t.Sequence], t.Optional[E]]:
-    def func(args: t.Sequence) -> t.Optional[E]:
-        if len(args) < 2:
-            return None
-
-        interval = args[1]
-        expression = interval.this
-        if expression and expression.is_string:
-            expression = exp.Literal.number(expression.this)
-
-        return expression_class(
-            this=args[0],
-            expression=expression,
-            unit=exp.Literal.string(interval.text("unit")),
-        )
-
-    return func
 
 
 def _date_add_sql(
@@ -168,8 +150,8 @@ class BigQuery(Dialect):
                 unit=exp.Literal.string(seq_get(args, 1).name),  # type: ignore
                 this=seq_get(args, 0),
             ),
-            "DATE_ADD": date_add(exp.DateAdd),
-            "DATETIME_ADD": date_add(exp.DatetimeAdd),
+            "DATE_ADD": parse_date_delta_with_interval(exp.DateAdd),
+            "DATETIME_ADD": parse_date_delta_with_interval(exp.DatetimeAdd),
             "DIV": lambda args: exp.IntDiv(this=seq_get(args, 0), expression=seq_get(args, 1)),
             "REGEXP_CONTAINS": exp.RegexpLike.from_arg_list,
             "REGEXP_EXTRACT": lambda args: exp.RegexpExtract(
@@ -181,12 +163,12 @@ class BigQuery(Dialect):
                 if re.compile(str(seq_get(args, 1))).groups == 1
                 else None,
             ),
-            "TIME_ADD": date_add(exp.TimeAdd),
-            "TIMESTAMP_ADD": date_add(exp.TimestampAdd),
-            "DATE_SUB": date_add(exp.DateSub),
-            "DATETIME_SUB": date_add(exp.DatetimeSub),
-            "TIME_SUB": date_add(exp.TimeSub),
-            "TIMESTAMP_SUB": date_add(exp.TimestampSub),
+            "TIME_ADD": parse_date_delta_with_interval(exp.TimeAdd),
+            "TIMESTAMP_ADD": parse_date_delta_with_interval(exp.TimestampAdd),
+            "DATE_SUB": parse_date_delta_with_interval(exp.DateSub),
+            "DATETIME_SUB": parse_date_delta_with_interval(exp.DatetimeSub),
+            "TIME_SUB": parse_date_delta_with_interval(exp.TimeSub),
+            "TIMESTAMP_SUB": parse_date_delta_with_interval(exp.TimestampSub),
             "PARSE_TIMESTAMP": lambda args: exp.StrToTime(
                 this=seq_get(args, 1), format=seq_get(args, 0)
             ),
